requirements.txt
flask
rembg[lite]
pillow
flasgger
